+++ src/components/checkbox/checkbox-base.ts
@@ -5,7 +5,7 @@ import { addKeyboardFocusRing } from '../common/controllers/focus-ring.js';
 import { blazorDeepImport } from '../common/decorators/blazorDeepImport.js';
 import type { Constructor } from '../common/mixins/constructor.js';
 import { EventEmitterMixin } from '../common/mixins/event-emitter.js';
-import { FormAssociatedRequiredMixin } from '../common/mixins/form-associated-required.js';
+import { FormAssociatedCheckboxRequiredMixin } from '../common/mixins/forms/associated-required.js';
 import { checkBoxValidators } from './validators.js';
 
 export interface CheckboxChangeEventArgs {
@@ -23,7 +23,7 @@ export interface IgcCheckboxEventMap {
 }
 
 @blazorDeepImport
-export class IgcCheckboxBaseComponent extends FormAssociatedRequiredMixin(
+export class IgcCheckboxBaseComponent extends FormAssociatedCheckboxRequiredMixin(
   EventEmitterMixin<IgcCheckboxEventMap, Constructor<LitElement>>(LitElement)
 ) {
   protected override get __validators() {
