+++ stories/checkbox.stories.ts
@@ -33,7 +33,8 @@ const metadata: Meta<IgcCheckboxComponent> = {
     },
     required: {
       type: 'boolean',
-      description: 'Makes the control a required field in a form context.',
+      description:
+        'When set, makes the component a required field for validation.',
       control: 'boolean',
       table: { defaultValue: { summary: false } },
     },
@@ -44,7 +45,7 @@ const metadata: Meta<IgcCheckboxComponent> = {
     },
     disabled: {
       type: 'boolean',
-      description: 'The disabled state of the component',
+      description: 'The disabled state of the component.',
       control: 'boolean',
       table: { defaultValue: { summary: false } },
     },
@@ -88,11 +89,11 @@ export default metadata;
 interface IgcCheckboxArgs {
   /** Draws the checkbox in indeterminate state. */
   indeterminate: boolean;
-  /** Makes the control a required field in a form context. */
+  /** When set, makes the component a required field for validation. */
   required: boolean;
   /** The name attribute of the control. */
   name: string;
-  /** The disabled state of the component */
+  /** The disabled state of the component. */
   disabled: boolean;
   /** Control the validity of the control. */
   invalid: boolean;
