+++ stories/date-time-input.stories.ts
@@ -73,7 +73,8 @@ const metadata: Meta<IgcDateTimeInputComponent> = {
     },
     required: {
       type: 'boolean',
-      description: 'Makes the control a required field in a form context.',
+      description:
+        'When set, makes the component a required field for validation.',
       control: 'boolean',
       table: { defaultValue: { summary: false } },
     },
@@ -84,7 +85,7 @@ const metadata: Meta<IgcDateTimeInputComponent> = {
     },
     disabled: {
       type: 'boolean',
-      description: 'The disabled state of the component',
+      description: 'The disabled state of the component.',
       control: 'boolean',
       table: { defaultValue: { summary: false } },
     },
@@ -150,11 +151,11 @@ interface IgcDateTimeInputArgs {
   locale: string;
   /** The prompt symbol to use for unfilled parts of the mask. */
   prompt: string;
-  /** Makes the control a required field in a form context. */
+  /** When set, makes the component a required field for validation. */
   required: boolean;
   /** The name attribute of the control. */
   name: string;
-  /** The disabled state of the component */
+  /** The disabled state of the component. */
   disabled: boolean;
   /** Control the validity of the control. */
   invalid: boolean;
