+++ stories/input.stories.ts
@@ -103,7 +103,8 @@ const metadata: Meta<IgcInputComponent> = {
     },
     required: {
       type: 'boolean',
-      description: 'Makes the control a required field in a form context.',
+      description:
+        'When set, makes the component a required field for validation.',
       control: 'boolean',
       table: { defaultValue: { summary: false } },
     },
@@ -114,7 +115,7 @@ const metadata: Meta<IgcInputComponent> = {
     },
     disabled: {
       type: 'boolean',
-      description: 'The disabled state of the component',
+      description: 'The disabled state of the component.',
       control: 'boolean',
       table: { defaultValue: { summary: false } },
     },
@@ -192,11 +193,11 @@ interface IgcInputArgs {
    * string-type inputs or allows spin buttons to exceed the predefined `min/max` limits for number-type inputs.
    */
   validateOnly: boolean;
-  /** Makes the control a required field in a form context. */
+  /** When set, makes the component a required field for validation. */
   required: boolean;
   /** The name attribute of the control. */
   name: string;
-  /** The disabled state of the component */
+  /** The disabled state of the component. */
   disabled: boolean;
   /** Control the validity of the control. */
   invalid: boolean;
