+++ stories/mask-input.stories.ts
@@ -59,7 +59,8 @@ const metadata: Meta<IgcMaskInputComponent> = {
     },
     required: {
       type: 'boolean',
-      description: 'Makes the control a required field in a form context.',
+      description:
+        'When set, makes the component a required field for validation.',
       control: 'boolean',
       table: { defaultValue: { summary: false } },
     },
@@ -70,7 +71,7 @@ const metadata: Meta<IgcMaskInputComponent> = {
     },
     disabled: {
       type: 'boolean',
-      description: 'The disabled state of the component',
+      description: 'The disabled state of the component.',
       control: 'boolean',
       table: { defaultValue: { summary: false } },
     },
@@ -133,11 +134,11 @@ interface IgcMaskInputArgs {
   mask: string;
   /** The prompt symbol to use for unfilled parts of the mask. */
   prompt: string;
-  /** Makes the control a required field in a form context. */
+  /** When set, makes the component a required field for validation. */
   required: boolean;
   /** The name attribute of the control. */
   name: string;
-  /** The disabled state of the component */
+  /** The disabled state of the component. */
   disabled: boolean;
   /** Control the validity of the control. */
   invalid: boolean;
