+++ stories/rating.stories.ts
@@ -95,7 +95,7 @@ const metadata: Meta<IgcRatingComponent> = {
     },
     disabled: {
       type: 'boolean',
-      description: 'The disabled state of the component',
+      description: 'The disabled state of the component.',
       control: 'boolean',
       table: { defaultValue: { summary: false } },
     },
@@ -156,7 +156,7 @@ interface IgcRatingArgs {
   allowReset: boolean;
   /** The name attribute of the control. */
   name: string;
-  /** The disabled state of the component */
+  /** The disabled state of the component. */
   disabled: boolean;
   /** Control the validity of the control. */
   invalid: boolean;
